chore: Import my earlier, field-tested `distcc.sh` wrapper
This original version of the script was designed for DistCC version 2.y
and uses manually configured SSH tunnels to detect connection availability
to the worker servers.
In addition, under the current system, the user had to manually
configure the number of jobs possible to be distributed to the target
machine as well.
Several beneficial features we would like to continue shipping, such as
automatic scaling and the handling of local job fallback in case no remote
machines are available are part of the code, however.

**This code is only injected here for reference.**
Our goal is to re-develop the clientside driver solution using the
existing infrastructure code only as a guardrail, but create a new
solution.

This is a combination of 5 commits.

- [distcc] Add DistCC package
    (Thu Aug 13 19:08:19 2020 +0200)
- [distcc] Allow the distcc script to scale workers based on available
  memory
    (Wed Aug 19 15:43:23 2020 +0200)
- [distcc] Fix DistCC_SH miscalculating local workers if cfg empty
  string
    (Sun Dec 13 12:39:16 2020 +0100)
- [distcc] Fix division by empty string error if DISTCC_LOCAL_MEM is
  unset
    (Sun Dec 13 12:51:41 2020 +0100)
- [distcc] Make listen socket detection pick up IPv6 sockets too
    (Thu Feb 11 17:31:32 2021 +0100)