wip: Start working on a `v2` version, add debugging infrastructure and start docs
The goal of the new version is to have direct TCP connection
specifications (just like raw `DISTCC_HOSTS`) and let go of the
**mandatory** nature of the _SSH Tunnels_ that was a requirement
previously, along with supporting the `--stats` feature of the
DistCC v3.y line.
This requires starting off with a more elaborate host specification
parsing routine, for which the basic layout (but no actual parsing!)
was laid down in this commit.