wip: Implement parsing of raw TCP hostspec
Supporting native TCP-based connections (in the same way as official
`DISTCC_HOSTS`) is the first major addition compared to `v1` of the script.

Users will be expected to provide the list of worker machines in the
format `hostname/ip-addr/ipv6-addr[:port[:stats]]`, which roughly
corresponds to `DISTCC_HOSTS="hostname:port/?"`, as we are preparing to
query the available job-count from the `stats` port automatically.