wip: Prepare for configuring "early" local jobs, scale them to RAM
In case a very small number of translation units (TUs) are recompiled
during an interactive compilation run, it is usually much faster to just
let the local machine deal with running that small number of compilers
instead of dealing with the distribution.
This patch reimplements `DISTCC_NUM_FIRST_LOCAL` as
`DISTCC_AUTO_EARLY_LOCAL_JOBS`, where the user might (although the
default is `"0"`) specify the number of purely local (non-distributed)
compiler invocations to prefer.

This value is scaled down automatically based on the available system
memory (as reported by `free`, which should be robust and real enough
for modern kernels) and the specified `DISTCC_AUTO_COMPILER_MEMORY`.
Thus, the previous `DISTCC_NUM_LOCAL_SCALING` is eliminated and does not
warrant reimplementation.